requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
black==24.10.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
coincurve==20.0.0
coverage==7.6.2
cytoolz==0.12.3
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozendict==2.3.10
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.0
parsimonious==0.10.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.26.1
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.3
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.24
tomli==2.0.2
toolz==0.12.1
tornado==6.4
tronpy==0.5.0
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
web3==7.3.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
